feat: support usernames through synthetic emails (#12)
firebase uses an email as the username, but scim doesn't have this
restriction.

we can workaround this by "synthesizing" a valid email from the
username, and then just being careful to perform this during login
(which tbf will be annoying as it'll have to happen client-side)

this is achieved by base36 encoding the userName and using this in front
of a staticly configured domain.

I haven't fully thought through if this is a good idea or not yet - eg:
I'm not sure how this will play out for OIDC / SAML

also includes a few misc fixes from testing with the Entra validator and
others.